fixed cell clearing bug
Insane python bug...

We have a loop in `sheet.set_cell_contents` which recomputes the extent
of the sheet after a cell is cleared. The loop uses an iterator
variable called `location` which collided with one of the function
arguments and modified it. This caused the cell being returned to
differ from the one whose contents was being changed.